+++ zerozstack-server-core/src/test/java/com/zeroz4j/server/ServerLiveMutexProviderTest.java
@@ -34,6 +34,7 @@ public class ServerLiveMutexProviderTest {
 
     @WeldSetup
     public WeldInitiator weld = WeldInitiator.of(
+            ServerRuntime.class,
             LiveMutexManager.class,
             ObjectMapperProducer.class
     );